Move MatfreeBlockDiag models to a different submodule
