+++ tests/test_probdiffeq/test_constraints/test_ts1_projected.py
@@ -56,13 +56,10 @@ def test_both_projected_constraints_are_identical(seed, num_probes):
 
     # Build the rest of the solver (projected, medium precision)
     key = random.prng_key(seed=seed)
-    ssm = probdiffeq.state_space_model_blockdiag()
+    ssm = probdiffeq.state_space_model_blockdiag_matfree(key=key, num_probes=num_probes)
     prior = ssm.prior_wiener_integrated(tcoeffs)
-    ode_ts1_projected = ssm.constraint_ode_ts1_projected(
-        vf, key=key, num_probes=num_probes
-    )
-
-    solver = probdiffeq.solver(strategy=strategy, constraint=ode_ts1_projected)
+    ode_ts1 = ssm.constraint_ode_ts1(vf)
+    solver = probdiffeq.solver(strategy=strategy, constraint=ode_ts1)
     solve = ivpsolve.solve_fixed_grid(solver=solver)
     solution_projected = solve(prior, grid=grid)
